# docs/api/crm-erp-rest-contract.openapi.yaml
openapi: 3.1.0
info:
  title: CRM/ERP Core Contract
  version: 1.0.0
  description: REST contract aligned to existing CRM/ERP schema and statuses.
servers:
  - url: /api/v1
paths:
...
  /orders/{orderId}/confirm:
    post:
      summary: Confirm order (atomic)
      operationId: confirmOrder
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmOrderRequest"
      responses:
        "200":
          description: Order confirmed with generated artifacts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfirmOrderResponse"
        "409":
          description: Business rule conflict (invalid status, insufficient stock, stale version)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
    OrderId:
      name: orderId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    ConfirmOrderRequest:
      type: object
      required: [expectedVersion, actorId, reservationPolicy]
      properties:
        expectedVersion:
          type: integer
          minimum: 1
        actorId:
          type: string
          format: uuid
        reservationPolicy:
          type: string
          enum: [FULL_ONLY, ALLOW_PARTIAL]
        warehouseId:
          type: string
          format: uuid
          description: Optional explicit warehouse for product reservations.
        correlationId:
          type: string
          format: uuid
    ConfirmOrderResponse:
      type: object
      required: [order, reservations, serviceRequests, events]
      properties:
        order:
          $ref: "#/components/schemas/Order"
        reservations:
          type: array
          items:
            $ref: "#/components/schemas/InventoryReservation"
        serviceRequests:
          type: array
          items:
            $ref: "#/components/schemas/ServiceRequest"
        events:
          type: array
          items:
            $ref: "#/components/schemas/EntityEventLog"
...
    Order:
      type: object
      required:
        - id
        - orderNo
        - customerId
        - status
        - currencyCode
        - subtotalAmount
        - discountAmount
        - taxAmount
        - totalAmount
        - version
      properties:
        id:
          type: string
          format: uuid
        orderNo:
          type: string
        customerId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/OrderStatus"
        orderDate:
          type: string
          format: date-time
        currencyCode:
          type: string
          pattern: "^[A-Z]{3}$"
        subtotalAmount:
          type: string
          description: Numeric(18,4) represented as string.
        discountAmount:
          type: string
        taxAmount:
          type: string
        totalAmount:
          type: string
        confirmedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        cancelledAt:
          type: string
          format: date-time
          nullable: true
        cancelReason:
          type: string
          nullable: true
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
...
    InventoryReservation:
      type: object
      required:
        - id
        - orderId
        - orderLineId
        - warehouseId
        - productId
        - status
        - reservedQty
        - releasedQty
        - consumedQty
        - outstandingQty
      properties:
        id:
          type: string
          format: uuid
        orderId:
          type: string
          format: uuid
        orderLineId:
          type: string
          format: uuid
        warehouseId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/InventoryReservationStatus"
        reservedQty:
          type: string
        releasedQty:
          type: string
        consumedQty:
          type: string
        outstandingQty:
          type: string
        reservedAt:
          type: string
          format: date-time
        releasedAt:
          type: string
          format: date-time
          nullable: true
        consumedAt:
          type: string
          format: date-time
          nullable: true
    ServiceRequest:
      type: object
      required:
        - id
        - requestNo
        - customerId
        - nameSnapshot
        - status
        - priority
      properties:
        id:
          type: string
          format: uuid
        requestNo:
          type: string
        orderId:
          type: string
          format: uuid
          nullable: true
        orderLineId:
          type: string
          format: uuid
          nullable: true
        customerId:
          type: string
          format: uuid
        catalogServiceId:
          type: string
          format: uuid
          nullable: true
        nameSnapshot:
          type: string
        status:
          $ref: "#/components/schemas/ServiceRequestStatus"
        priority:
          $ref: "#/components/schemas/ServiceRequestPriority"
        plannedStartAt:
          type: string
          format: date-time
          nullable: true
        plannedEndAt:
          type: string
          format: date-time
          nullable: true
        actualStartAt:
          type: string
          format: date-time
          nullable: true
        actualEndAt:
          type: string
          format: date-time
          nullable: true
        assigneeId:
          type: string
          format: uuid
          nullable: true
        slaDueAt:
          type: string
          format: date-time
          nullable: true
        resolutionNote:
          type: string
          nullable: true
    EntityEventLog:
      type: object
      required:
        - id
        - entityType
        - entityId
        - eventType
        - operation
        - payload
        - occurredAt
      properties:
        id:
          type: string
          format: uuid
        businessId:
          type: string
          format: uuid
          nullable: true
        entityType:
          type: string
        entityId:
          type: string
          format: uuid
        eventType:
          type: string
        operation:
          $ref: "#/components/schemas/AuditOperation"
        orderId:
          type: string
          format: uuid
          nullable: true
        orderLineId:
          type: string
          format: uuid
          nullable: true
        serviceRequestId:
          type: string
          format: uuid
          nullable: true
        reservationId:
          type: string
          format: uuid
          nullable: true
        movementId:
          type: string
          format: uuid
          nullable: true
        actorId:
          type: string
          format: uuid
        correlationId:
          type: string
          format: uuid
          nullable: true
        payload:
          type: object
          additionalProperties: true
        occurredAt:
          type: string
          format: date-time
        source:
          type: string
          nullable: true
...
    OrderStatus:
      type: string
      enum:
        - DRAFT
        - READY_FOR_CONFIRMATION
        - CONFIRMED
        - PARTIALLY_FULFILLED
        - FULFILLED
        - COMPLETED
        - CANCELLED
...
    ServiceRequestStatus:
      type: string
      enum:
        - NEW
        - PLANNED
        - ASSIGNED
        - IN_PROGRESS
        - DONE
        - FAILED
        - CANCELLED
    ServiceRequestPriority:
      type: string
      enum:
        - LOW
        - NORMAL
        - HIGH
        - URGENT
    InventoryReservationStatus:
      type: string
      enum:
        - ACTIVE
        - PARTIALLY_RELEASED
        - RELEASED
        - CONSUMED
        - CANCELLED
...
    AuditOperation:
      type: string
      enum:
        - INSERT
        - UPDATE
        - DELETE